+++ sdk/messaging/azservicebus/messageBatch.go
@@ -39,7 +39,7 @@ func newMessageBatch(maxBytes int) *MessageBatch {
 // (true, nil) if the message was added.
 // (false, nil) if the message was too large to fit into the batch.
 // (false, err) if an error occurs when adding the message.
-func (mb *MessageBatch) Add(m *Message) (bool, error) {
+func (mb *MessageBatch) Add(m SendableMessage) (bool, error) {
 	msg := m.toAMQPMessage()
 
 	if msg.Properties.MessageID == nil || msg.Properties.MessageID == "" {