+++ sdk/messaging/azservicebus/CHANGELOG.md
@@ -1,8 +1,14 @@
 # Release History
 
-## 0.1.1 (Unreleased)
+## 0.2.0 (Unreleased)
 
 ### Features Added
+
+- Scheduling messages to be delivered at a later date, via the `Sender.ScheduleMessage(s)` function or 
+  setting `Message.ScheduledEnqueueTime`.
+- Added in the `Sender.SendMessages([slice of sendable messages])` function, which batches messages 
+  automatically. Useful when you're sending multiple messages that you are already sure will be small
+  enough to fit into a single batch.
 
 ### Breaking Changes